Move triangle task to appropriate folder And added new solution of Task - Square
